xpath: Remove duplicates after sorting
Simply stop checking for duplicates when adding to intermediate results
and remove duplicates in a single pass after sorting.

This fixes #130 but can cause intermediate results to grow much too
large. This can also lead to a huge amount of unnecessary work for later
operations until the result is sorted.

This will be addressed in a later commit.