xpath: Remove duplicates from intermediate step results
The initial fix for #130 delayed duplicate removal from intermediate
step results for too long, leading to quadratic growth of duplicate
nodes in certain cases.

Now we deduplicate intermediate results dynamically if certain
thresholds are reached.

We still allow duplicate nodes to be passed to consecutive steps or
union operations, resulting in unnecessary work. But duplicates are
limited to 50% of a nodeset and should be rare in practice, so I think
this is a workable approach.

Note that after previous commits, the results of unions are also
deduplicated lazily which could slow down expressions like (a|b|c)/d.
But this should be acceptable as well.

If it doesn't work out, we must require input to step ops being
deduplicated.